FIX: Remove invalid bigint cast on trade_id in SQL queries
trade_id values are strings (e.g., 'TRD-C00F10875E'), not numeric bigints.
Casting to bigint caused 'InvalidTextRepresentation' SQL errors that aborted
transactions and cascaded to all subsequent exit checks, causing 'degraded'
status with '2 errors'.

Fixed in 3 locations:
- executor_exit_handler.py:534 (primary exit flow)
- reconciliation.py:1606 (partial PnL lookup)
- reconciliation.py:1824 (audit log lookup)

This resolves the root cause of exit execution failures and transaction aborts.

+++ algo/infrastructure/reconciliation.py
@@ -1603,7 +1603,7 @@ class DailyReconciliation:
                         SELECT COALESCE(SUM((details->>'pnl_dollars')::numeric), 0)
                         FROM algo_audit_log
                         WHERE action_type LIKE 'exit_%%'
-                          AND (details->>'trade_id')::bigint = %s
+                          AND (details->>'trade_id') = %s
                           AND (details->>'full_exit')::boolean = false
                         """,
                         (trade_id,),
@@ -1821,7 +1821,7 @@ class DailyReconciliation:
                        COALESCE(SUM((details->>'shares_exited')::numeric), 0)
                 FROM algo_audit_log
                 WHERE action_type LIKE 'exit_%%'
-                  AND (details->>'trade_id')::bigint = %s
+                  AND (details->>'trade_id') = %s
                   AND (details->>'full_exit')::boolean = false
                 """,
                 (trade_id,),
